share a common N5PathSupplier
