+++ src/main/java/org/janelia/saalfeldlab/hotknife/SparkComputeCostBrainVNC.java
@@ -10,7 +10,7 @@ import java.util.concurrent.Executors;
 import org.apache.spark.SparkConf;
 import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.api.java.JavaSparkContext;
-import org.janelia.saalfeldlab.hotknife.SparkComputeCost.N5PathSupplier;
+import org.janelia.saalfeldlab.hotknife.util.N5PathSupplier;
 import org.janelia.saalfeldlab.hotknife.util.Show;
 import org.janelia.saalfeldlab.n5.DataType;
 import org.janelia.saalfeldlab.n5.GzipCompression;
